requirements.txt
PyQt5==5.15.2
aniso8601==9.0.1
asn1crypto==1.5.1
black==22.10.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.3
coverage==7.3.2
Faker==19.10.0
Flask==3.0.0
Flask-Cors==3.0.10
Flask-JWT-Extended==4.5.3
Flask-Migrate==4.0.5
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-Testing==0.8.1
greenlet==3.0.0
gunicorn==21.2.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
mypy-extensions==0.4.3
pathspec==0.10.1
pg8000==1.29.3
psycopg2-binary==2.9.9
PyJWT==2.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
responses==0.23.3
scramp==1.4.4
six==1.16.0
SQLAlchemy==2.0.21
types-PyYAML==6.0.12.11
typing_extensions==4.4.0
urllib3==2.0.6
Werkzeug==3.0.0
